requirements.txt
gphoto2
pyserial
rawpy
pillow
tqdm